feat(pt2): Display a list
The Tour of Heroes application displays a list of heroes with a detail view.
The user can select a hero and see that hero's details.
You used *ngFor to display a list.
You used *ngIf to conditionally include or exclude a block of HTML.
You can toggle a CSS style class with a class binding.